WB5: boundary strips and root-file moves
