+++ src/ui/remoteSource/index.ts
@@ -1,0 +1,9 @@
+export type { CompanionAssetSummary } from './sessionAssets.svelte';
+export {
+	companionSummaryForInputIds,
+	hasSupplementalAssetsForInputId,
+	purgeRemoteSourceSessionsForInputIds,
+	releaseRemoteSourceSessionRetainers,
+	retainRemoteSourceSessionsForInputIds,
+	registerRemoteSourceSupplementalAssets,
+} from './sessionAssets.svelte';
